Ajout du widget ModernGradientButton et mise à jour des écrans
