+++ lib/screens/home_screen.dart
@@ -5,6 +5,7 @@ import 'cadre_list_screen.dart';
 import 'search_screen.dart';
 import 'quiz_menu_screen.dart';
 import '../widgets/ad_banner.dart';
+import '../widgets/modern_gradient_button.dart';
 
 class HomeScreen extends StatelessWidget {
   const HomeScreen({Key? key}) : super(key: key);
@@ -38,7 +39,7 @@ class HomeScreen extends StatelessWidget {
                     children: [
                       FractionallySizedBox(
                         widthFactor: 0.85,
-                        child: _ModernGradientButton(
+                        child: ModernGradientButton(
                           icon: Icons.book,
                           label: 'Infractions par thèmes',
                   onPressed: () {
@@ -56,7 +57,7 @@ class HomeScreen extends StatelessWidget {
               const SizedBox(height: 16),
               FractionallySizedBox(
                 widthFactor: 0.85,
-                child: _ModernGradientButton(
+                child: ModernGradientButton(
                   icon: Icons.gavel,
                   label: "Cadres d'enquête",
                   onPressed: () {
@@ -74,7 +75,7 @@ class HomeScreen extends StatelessWidget {
               const SizedBox(height: 16),
               FractionallySizedBox(
                 widthFactor: 0.85,
-                child: _ModernGradientButton(
+                child: ModernGradientButton(
                   icon: Icons.search,
                   label: "Rechercher une infraction",
                   onPressed: () {
@@ -92,7 +93,7 @@ class HomeScreen extends StatelessWidget {
               const SizedBox(height: 16),
               FractionallySizedBox(
                 widthFactor: 0.85,
-                child: _ModernGradientButton(
+                child: ModernGradientButton(
                   icon: Icons.star,
                   label: 'Mes favoris',
                   onPressed: () {
@@ -110,7 +111,7 @@ class HomeScreen extends StatelessWidget {
               const SizedBox(height: 16),
               FractionallySizedBox(
                 widthFactor: 0.85,
-                child: _ModernGradientButton(
+                child: ModernGradientButton(
                   icon: Icons.quiz,
                   label: 'Quiz',
                   onPressed: () {
@@ -128,98 +129,6 @@ class HomeScreen extends StatelessWidget {
               const SizedBox(height: 32),
               const AdBanner(),
             ],
-          ),
-        ),
-      ),
-    ],
-    ),
-    ),
-      )
-      );
-  }
-}
-
-// Widget bouton moderne avec dégradé subtil, coins peu arrondis, effet hover/clic
-class _ModernGradientButton extends StatefulWidget {
-  final IconData icon;
-  final String label;
-  final VoidCallback onPressed;
-  const _ModernGradientButton({
-    required this.icon,
-    required this.label,
-    required this.onPressed,
-    Key? key,
-  }) : super(key: key);
-
-  @override
-  State<_ModernGradientButton> createState() => _ModernGradientButtonState();
-}
-
-class _ModernGradientButtonState extends State<_ModernGradientButton> {
-  bool _hovering = false;
-  bool _pressed = false;
-
-  void _setHover(bool value) {
-    setState(() {
-      _hovering = value;
-    });
-  }
-
-  void _setPressed(bool value) {
-    setState(() {
-      _pressed = value;
-    });
-  }
-
-  @override
-  Widget build(BuildContext context) {
-    double opacity = 1.0;
-    if (_pressed) {
-      opacity = 0.82;
-    } else if (_hovering) {
-      opacity = 0.90;
-    }
-    return MouseRegion(
-      onEnter: (_) => _setHover(true),
-      onExit: (_) => _setHover(false),
-      child: GestureDetector(
-        onTapDown: (_) => _setPressed(true),
-        onTapUp: (_) => _setPressed(false),
-        onTapCancel: () => _setPressed(false),
-        child: AnimatedOpacity(
-          duration: const Duration(milliseconds: 120),
-          opacity: opacity,
-          child: Container(
-            decoration: BoxDecoration(
-              gradient: const LinearGradient(
-                colors: [Color(0xFFFAF9F6), Colors.white,Color(0xFFFAF9F6)],
-                begin: Alignment.topLeft,
-                end: Alignment.bottomRight,
-              ),
-              borderRadius: BorderRadius.circular(8),
-              boxShadow: const [
-                BoxShadow(
-                  color: Colors.black26,
-                  blurRadius: 8,
-                  offset: Offset(0, 4),
-                ),
-              ],
-            ),
-            child: ElevatedButton.icon(
-              icon: Icon(widget.icon),
-              label: Text(widget.label),
-              style: ElevatedButton.styleFrom(
-                backgroundColor: Colors.transparent,
-                shadowColor: Colors.transparent,
-                shape: RoundedRectangleBorder(
-                  borderRadius: BorderRadius.circular(8),
-                ),
-                padding: const EdgeInsets.symmetric(horizontal: 24, vertical: 16),
-                textStyle: const TextStyle(fontSize: 18, fontWeight: FontWeight.w600),
-                elevation: 0,
-              ),
-              onPressed: widget.onPressed,
-            ),
           ),
         ),
       ),
